Cleanup after file moving
