+++ sandbox_storage/dao/db_models.py
@@ -18,7 +18,7 @@ Database Models
 """
 
 from sqlalchemy import Column, Integer, String, DateTime
-from .database import Base
+from .db import Base
 
 
 class DrsObject(Base):
